+++ test/hash.cpp
@@ -11,7 +11,7 @@ using namespace fur::hash;
 
 static void check_string(const std::string& string, const hash_t& hash) {
   REQUIRE(string.length() == 20);
-  for (int i = 0; i < 10; i++) {
+  for (int64_t i = 0; i < 10; i++) {
     REQUIRE(static_cast<uint8_t>(string[i]) == hash[i]);
   }
 }
@@ -50,8 +50,8 @@ TEST_CASE("[Hash] Split pieces string") {
   auto r = split_piece_hashes(std::string{piece_hashes, 3 * 20});
   REQUIRE(r.valid());
   auto pieces = *r;
-  REQUIRE((pieces).size() == 3);
-  for (int i = 0; i < 3; i++) {
+  REQUIRE(pieces.size() == 3);
+  for (int64_t i = 0; i < 3; i++) {
     hash_t want;
     std::fill(want.begin(), want.end(), i);
     REQUIRE(pieces[i] == want);
